Updates after code review
* remove dead condition - due to nature of importer
* minor fixes
